Split yaml_c_wrapper.cpp into logically organized files
The single yaml_c_wrapper.cpp mixed the generic YAML tree wrapper with all
the PALS-specific lattice logic. Split it along its natural seams:

  - yaml_c_wrapper.cpp  generic YAML tree wrapper (ryml C API): parse,
                        traverse, query, modify, emit. Knows nothing about PALS.
  - pals_expand.cpp     the four-tree expansion pipeline (original / combined /
                        expanded / leftover): include splicing, structural
                        expansion (repeats, inherits, forks), and expression /
                        controller evaluation.
  - pals_match.cpp      PALS name matching and parameter lookup (the only
                        PCRE2 code).
  - pals_util.{h,cpp}   the few helpers that straddle the expand/match seam:
                        child_val_str, split_dots, resolve_param_path,
                        strip_expr_wrapper.
  - yaml_tree.h         the shared tree representation (ParsedData, GET_TREE,
                        ensure_capacity, deep_copy_recursive) used by both the
                        generic wrapper and the PALS code.

Pure code movement: no behavior change. The public C API in yaml_c_wrapper.h
is unchanged; all 104 tests pass.
